fix ts task
